+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.0.41
+
+* Add `DropArea`
+* Add `UIKitPlaceholder`
+* `UIKitForm#Custom` now have nullable param target
+* Add `UIKitUtility#Link`
+* Add `DescriptionList` and several support composable functions
+
 ## 0.0.40
 
 * All `DefaultTable` functions now use `Iterable` as data type